Partial fix for millisecond timestamps where more than 32,767 data points were stored in a single RowSeq. Larry Reeder found this bug where an assertion was thrown because the short index rolled over. Changing it to an int will help but only allows up to 2.1M data points in a row whereas the schema allows 4M. TODO - figure out an efficient means of keeping the full 4M bytes in memory with quick access for iteration.

+++ src/core/RowSeq.java
@@ -509,10 +509,15 @@ final class RowSeq implements DataPoints {
     private int qualifier;
 
     /** Next index in {@link #qualifiers}.  */
-    private short qual_index;
+    // TODO - This was a short, which was fine for the second qualifiers but
+    // now with ms support we can have up to 2^22 = 4194304 values in a row. 
+    // Changing to an int helps a little but will rollover at 2,147,483,647 at 
+    // which point we can't reference the array. We need to redo the RowSeq 
+    // storage so it can be referenced above 2.1M data points
+    private int qual_index;
 
     /** Next index in {@link #values}.  */
-    private short value_index;
+    private int value_index;
 
     /** Pre-extracted base time of this row sequence.  */
     private final long base_time = baseTime();
